+++ fluxo-kmp-conf/src/main/kotlin/FkcSetupGradlePlugin.kt
@@ -4,7 +4,7 @@
 
 import fluxo.conf.dsl.FluxoConfigurationExtension
 import fluxo.conf.dsl.fluxoConfiguration
-import fluxo.conf.feat.gradlePluginExt
+import fluxo.conf.pub.gradlePluginExt
 import fluxo.log.e
 import fluxo.log.l
 import fluxo.log.w