+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/FluxoConfigurationExtensionKotlin.kt
@@ -136,7 +136,7 @@ public interface FluxoConfigurationExtensionKotlin : FluxoConfigurationExtension
      * Inherited from the parent project if not set.
      * Default value: `false`.
      */
-    public var useDokka: Boolean?
+    public var useDokka: Boolean
 
 
     // FIXME: koverReport settings