+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/dsl/FluxoPublicationConfig.kt
@@ -1,5 +1,6 @@
 package fluxo.conf.dsl
 
+import com.vanniktech.maven.publish.SonatypeHost
 import fluxo.conf.dsl.FluxoConfigurationExtensionPublication.Companion.DEFAULT_BRANCH_NAME
 
 // TODO: Make an immutable resulting class
@@ -10,6 +11,8 @@ public data class FluxoPublicationConfig(
     public var version: String,
     /** Artifact ID, basically */
     public var projectName: String? = null,
+
+    public var inceptionYear: String? = null,
 
     public var projectDescription: String? = null,
     /** Project website or repository url */
@@ -23,6 +26,8 @@ public data class FluxoPublicationConfig(
 
     /** PGP signing key */
     public var signingKey: String? = null,
+    /** PGP signing key ID */
+    public var signingKeyId: String? = null,
     /** PGP signing password */
     public var signingPassword: String? = null,
 
@@ -31,13 +36,17 @@ public data class FluxoPublicationConfig(
     /** Maven publishing repository password */
     public var repositoryPassword: String? = null,
 
-    /** VCS link to the specific publication (e.g. release tag) */
+    /** VCS link to the specific publication (e.g., release tag) */
     public var publicationUrl: String? = projectUrl,
     public var isSnapshot: Boolean = version.contains("SNAPSHOT", ignoreCase = true),
 
     /** VCS tag or branch name for publication */
     public var scmTag: String? = if (isSnapshot) DEFAULT_BRANCH_NAME else "v$version",
 
+    /**
+     * Use instead of [repositoryUrl] for the Vanniktech's Maven Publish plugin.
+     */
+    public var sonatypeHost: SonatypeHost = SonatypeHost.DEFAULT,
     public var repositoryUrl: String = when {
         isSnapshot -> "https://s01.oss.sonatype.org/content/repositories/snapshots/"
         else -> "https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/"