+++ fluxo-kmp-conf/src/main/kotlin/fluxo/conf/pub/SetupGradlePublishPlugin.kt
@@ -1,10 +1,11 @@
-package fluxo.conf.feat
+package fluxo.conf.pub
 
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_ALIAS
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_ID
 import fluxo.conf.data.BuildConstants.GRADLE_PLUGIN_PUBLISH_PLUGIN_VERSION
 import fluxo.conf.deps.loadAndApplyPluginIfNotApplied
 import fluxo.conf.dsl.impl.FluxoConfigurationExtensionImpl
+import fluxo.conf.feat.setupValidatePluginTasks
 import org.gradle.api.Project
 
 internal fun Project.setupGradlePublishPlugin(conf: FluxoConfigurationExtensionImpl) {
@@ -15,7 +16,8 @@ internal fun Project.setupGradlePublishPlugin(conf: FluxoConfigurationExtensionI
         project = this,
 
         // Lookup will fail with AmbiguousGraphVariantsException
-        // (shadowRuntimeElements variant is needed, but not resolved dynamically at this moment)
+        // (shadowRuntimeElements variant is needed
+        // but not resolved dynamically at this moment)
         lookupClassName = false,
 //        className = GRADLE_PLUGIN_PUBLISH_CLASS_NAME,
     )
@@ -24,4 +26,5 @@ internal fun Project.setupGradlePublishPlugin(conf: FluxoConfigurationExtensionI
 }
 
 @Suppress("unused")
+// language=jvm-class-name
 private const val GRADLE_PLUGIN_PUBLISH_CLASS_NAME = "com.gradle.publish.PublishPlugin"