+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - fix bundled shrinker loading.
 
 ### Changed
+- use `vanniktech/gradle-maven-publish-plugin` for publication.
 - migrate to the new Develocity plugin for the build scans.
 - prefer bundled R8 if it’s newer by default.
 - change default values and logic for some configuration options for safety and convenience in the new projects (experimental and/or complicated options should be opt-in)